requirements.txt
google-genai
google-adk
pandas
python-dotenv
mcp
kagglehub